Initial revamp on psa
Templates still reside in {prefix}/share/psa/templates, but
will have a template.cfg file inside that will describe the
template. Curently only the template name and the command
class to be used. These command classes are abstraction
of the commands to be executed to work with a psa-based
project. Implemented classes are DebProject, Fremantle
and Harmattan.

Also when creating a project a {slug}.psa file is created
with the info about the project as a mean to load that
information again if needed during the other commands.

Other changes:

* Instead of passing the argument parser around, call
  it only if a wrong command was given.
* Removed unused arguments to some functions.
* Will show the list of available platforms if an
  unavailable platform is chosen.
* Smaller psa_init method. Wrapped some code into smaller
  functions.
* Smaller changes to make pep8 and pylint happier
* README now add .assistant suffix only on created
  projects.